bai14

+++ app/Http/Controllers/Backend/PostController.php
@@ -7,9 +7,23 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Models\Tag;
+use Illuminate\Support\Facades\Gate;
+// use App\Http\Controllers\Auth;
+use Illuminate\Support\Facades\Auth;
+
+
+
 
 class PostController extends Controller
 {
+    /**
+     * Create the controller instance.
+     * 
+     * @return void
+     */
+    public function _construct() {
+        $this->authorizeResource(Post::class,'post');
+    }    
     /**
      * Display a listing of the resource.
      *
@@ -18,7 +32,7 @@ class PostController extends Controller
     public function index(Request $request)
     {
         //
-        $users = Post::all();
+        // $users = Post::all();
         // dd($users);
         $title = request()->get('title');
         $status = request()->get('status');
@@ -30,9 +44,6 @@ class PostController extends Controller
             $posts_query = $posts_query->where('status',$status);
         }
         $posts = $posts_query->paginate(3);
-        // dd($posts);
-        // $posts = DB::table('posts')->get();
-        // dd($posts);
         return view('backend.posts.index')->with([
             'posts' => $posts
         ]);;
@@ -62,6 +73,7 @@ class PostController extends Controller
         //     'title','content'
         // ]);
         $data = $request->all();
+        
         // dd($data);
         $post = new Post();
         $post->title = $data['title'];
@@ -71,33 +83,8 @@ class PostController extends Controller
         $post->category_id = 1;
         $post->content = $data['content'];
         $post->save();
-        // DB::table('posts')->insert([
-        //     'title' => $data['title'],
-        //     'slug' => Str::slug($data['title']),
-        //     'content' => $data['content'],
-        //     'user_created_id' => 1,
-        //     'user_updated_id' => 1,
-        //     'category_id' => 1,
-        //     'created_at' => now(),
-        //     'updated_at' => now()
-
-        // ]);
-
         return redirect()->route('backend.posts.index');
-    //     $data= true;
-    //     if($data) {
-    //             return redirect()->route('backend.posts.index');
-    //             // return redirect()->action([PostController::class , 'index']);
-    //     }else {
-    //         return redirect()->back();
-    //     }
-    //    if($request->is('backend/*')) {
-    //        dd('dung');
-    //    }else {
-    //         dd('error');
-    //    };
-    //    dd($request->path());
-    //    dd($request->except('_token'));
+    
     }
 
     /**
@@ -119,14 +106,6 @@ class PostController extends Controller
         foreach($post->tags as $tag) {
             echo $tag->name;
         }   ;
-        // dd($post);
-        // if($id) {
-        //     return view('backend.posts.index');
-        // }else {
-        //     return 'dungh';
-        // }
-        
-
     }
 
     /**
@@ -142,9 +121,12 @@ class PostController extends Controller
         if($id !== null) {
             // $post = Post::firsWhere('id', $id);
             // dd($post);
-            $posts = DB::table('posts')->select(['id','title','content'])->find($id);
+            // $posts = DB::table('posts')->select(['id','title','content'])->find($id);
+            $posts = Post::find($id);
+            $tags = Tag::find($id);
             return view('Backend.posts.edit')->with([
-                'posts'=>$posts
+                'posts'=>$posts,
+                'tags'=>$tags,
             ]);
            
         }else {
@@ -161,13 +143,24 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request,Post $post)
     {
         //
-        $data = $request->all();
-        $post = Post::find($id);
+        // $post = Post::find($id);
+        // dd($post);
+        // if (! Gate::allows('update-post', $post)) {
+        //     abort(403);
+        // };
+        // if($request->user()->cannot('update',$post)) {
+        //     abort(403);
+        // }
+        $this->authorize('update', $post);
+        $data = $request->only(['title','content']);
+        $data = $request->tag;
         $post->title = $data['title'];
         $post->content = $data['content'];
+        $post->user_id = Auth::user()->id;
+        // dd($post);
         $post->save();
         return redirect()->route('backend.posts.index');
     }
@@ -181,6 +174,10 @@ class PostController extends Controller
     public function destroy($id)
     {
         //
+        $post = Post::find($id);
+        if (! Gate::allows('delete-post', $post)) {
+            abort(403);
+        };
         $post = Post::find($id);
         $post->delete();
         return redirect()->route('backend.posts.index');
